requirements.txt
streamlit==1.32.0
pydantic==2.5.0
sqlmodel==0.0.14
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pypdf==4.0.1
python-docx==0.8.11
python-pptx==0.6.23
python-dotenv==1.0.0
google-generativeai==0.3.0

